refactor(wallet/ui): factor out chainInfo.js
